Former-commit-id: 8182bb8e4ac791643076a4e3320e6e9df330ed07

+++ tools/update_env_var.sh
@@ -3,6 +3,8 @@
 var=$1
 value=$2
 
-cat ps_run.sh |sed "s/$1=.*$/$1=$2/" > ps_run.sav
-mv ps_run.sav ps_run.sh
-chmod +x ps_run.sh
+if [ -f ps_run.sh ]; then
+	cat ps_run.sh |sed "s/$1=.*$/$1=$2/" > ps_run.sav
+	mv ps_run.sav ps_run.sh
+	chmod +x ps_run.sh
+fi
